requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0
azure-search-documents>=11.6.0b5